"Refactor backend code: moved routes to separate blueprints, updated imports, and modified route registrations"
